feat: add raw data buttons
