+++ packages/round-manager/src/features/round/ViewRoundResults/ViewRoundResults.tsx
@@ -28,6 +28,7 @@ import { errorModalDelayMs } from "../../../constants";
 import { useRoundById } from "../../../context/round/RoundContext";
 import { TransactionResponse } from "@ethersproject/providers";
 import { payoutTokens } from "../../api/utils";
+import { roundApplicationsToCSV } from "../../api/exports";
 
 type RevisedMatch = Match & {
   revisedMatch: bigint;
@@ -228,11 +229,11 @@ export default function ViewRoundResults() {
     },
   ];
 
-  if (isBeforeRoundEndDate && !debugModeEnabled) {
+  if (!chain || (isBeforeRoundEndDate && !debugModeEnabled)) {
     return <NoInformationContent />;
   }
 
-  if (isLoadingRound) {
+  if (isLoadingRound || !round) {
     return <Spinner text="We're fetching the matching data." />;
   }
 
@@ -571,7 +572,55 @@ export default function ViewRoundResults() {
             </div>
           </Tab.Panel>
           <Tab.Panel>
-            <div>{/* raw round data content here */}</div>
+            <div className="text-sm leading-5 text-gray-400 text-left pb-4">
+              Download and use the data models provided to analyze your round
+              results in-depth.
+            </div>
+            <div className="text-sm leading-5 text-gray-500 font-semibold text-left mb-3 mt-2">
+              Round Generated Data
+            </div>
+            <div className="text-sm leading-5 text-gray-400 text-left">
+              Download the raw data for your round, which is separated into four
+              data tables: raw votes, projects, round, and prices.
+            </div>
+            <div className="pt-6">
+              <a
+                className="flex items-center mb-4"
+                href={`${process.env.REACT_APP_ALLO_API_URL}/api/v1/chains/${chain?.id}/rounds/${roundId}/exports/votes`}
+              >
+                <span className="w-40">Raw Votes</span>
+                <DownloadIcon className="h-5 w-5" />
+              </a>
+              <a
+                className="flex items-center mb-4"
+                href="#"
+                onClick={(e) => {
+                  e.preventDefault();
+                  exportAndDownloadApplicationsCSV(
+                    round.id,
+                    chain.id,
+                    chain.name
+                  );
+                }}
+              >
+                <span className="w-40">Projects</span>
+                <DownloadIcon className="h-5 w-5" />
+              </a>
+              <a
+                className="flex items-center mb-4"
+                href={`${process.env.REACT_APP_ALLO_API_URL}/api/v1/chains/${chain?.id}/rounds/${roundId}/exports/round`}
+              >
+                <span className="w-40">Round</span>
+                <DownloadIcon className="h-5 w-5" />
+              </a>
+              <a
+                className="flex items-center mb-4"
+                href={`${process.env.REACT_APP_ALLO_API_URL}/api/v1/chains/${chain?.id}/rounds/${roundId}/exports/prices`}
+              >
+                <span className="w-40">Prices</span>
+                <DownloadIcon className="h-5 w-5" />
+              </a>
+            </div>
           </Tab.Panel>
         </Tab.Panels>
       </Tab.Group>
@@ -720,4 +769,18 @@ export function downloadFile(data: BlobPart, filename: string): void {
   document.body.appendChild(link);
   link.click();
   document.body.removeChild(link);
+}
+
+async function exportAndDownloadApplicationsCSV(
+  roundId: string,
+  chainId: number,
+  chainName: string
+) {
+  const csv = await roundApplicationsToCSV(roundId, chainId, chainName);
+  // create a download link and click it
+  const blob = new Blob([csv], {
+    type: "text/csv;charset=utf-8;",
+  });
+
+  downloadFile(blob, `projects-${roundId}.csv`);
 }
